Move HTTP tests into corresponding folders
